add texture for electric essence and base staff, add config

+++ src/main/java/de/zekro/magicstaffs/util/Vec3dUtils.java
@@ -25,7 +25,7 @@ public class Vec3dUtils {
      * @param by multiply by
      * @return result
      */
-    public static Vec3d multiply(Vec3d vec, double by) {
+    public static Vec3d multiply(Vec3d vec, float by) {
         return multiply(vec, new Vec3d(by, by, by));
     }
 }
